requirements.txt
python-telegram-bot>=20.7
openai>=1.0.0
PyYAML>=6.0.1
pydantic-settings>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
